+++ EssentialFeedJorgeiOS/FeedUI/Controllers/FeedViewController.swift
@@ -10,19 +10,6 @@ import UIKit
 import EssentialFeedJorge
 
 
-// MARK: - Protocols
-public protocol FeedImageDataLoaderTask {
-    func cancel()
-}
-
-
-public protocol FeedImageDataLoaderProtocol {
-    typealias Result = Swift.Result<Data, Error>
-    
-    func loadImageData(from url: URL, completion: @escaping (Result) -> Void) -> FeedImageDataLoaderTask
-}
-
-
 // MARK: - FeedViewController Class
 public final class FeedViewController: UITableViewController  {
     private var feedLoader: FeedLoaderProtocol?
